+++ src/antithesis/_location.py
@@ -5,30 +5,30 @@ source code metadata for assertion callers.
 
 """
 
-from typing import Union, Any, Dict, Optional
-from inspect import FrameInfo
+from types import FrameType
+from typing import Union, Dict, Optional
 
 
-def _get_class_name(frame_info: Any) -> str:
+def _get_class_name(frame: FrameType) -> str:
     try:
-        class_name = frame_info.f_locals["self"].__class__.__name__
+        class_name = frame.f_locals["self"].__class__.__name__
     except KeyError:
         class_name = ""
     return class_name
 
 
-def _get_location_info(frame_info: Optional[FrameInfo]) -> Dict[str, Union[str, int]]:
+def _get_location_info(frame: Optional[FrameType]) -> Dict[str, Union[str, int]]:
     """Provides a dictionary containing source code info obtained from assertion callers.
 
     Args:
-        frame_info (:obj:`FrameInfo`, optional): Assertion caller's stack frame info or None
+        frame (:obj:`FrameType`, optional): Assertion caller's stack frame or None
 
     Returns:
         Dict[str, Union[str, int]]: a dictionary containing source code info
             obtained from assertion callers.
 
     """
-    if frame_info is None:
+    if frame is None:
         print("LocInfo not available")
         return {
             "file": "",
@@ -38,9 +38,9 @@ def _get_location_info(frame_info: Optional[FrameInfo]) -> Dict[str, Union[str, 
             "begin_column": 0,
         }
     return {
-        "file": frame_info.filename,
-        "function": frame_info.function,
-        "class": _get_class_name(frame_info.frame),
-        "begin_line": frame_info.lineno,
+        "file": frame.f_code.co_filename,
+        "function": frame.f_code.co_name,
+        "class": _get_class_name(frame),
+        "begin_line": frame.f_lineno,
         "begin_column": 0,
     }